GUI becoming complex. Reinstated examples/example_simple_GUI.py, moved new code to example_GUI.py
